Fix reading of the SESSION param from default file
When SESSION param was defined in the default config file
(/etc/extagent/default/) it did not behave correctly when starting the
extagent. Basically it was not possible to define only part of the
session string. It was always necessary to specify all params even if
they did not differ from the default values.

This also set the SLEEP time to zero in order to speed up the start time
of the extagent.